+++ src/configs/twitter.js
@@ -1,5 +1,16 @@
 module.exports = {
   twitter: {
     enabled: false,
+    api: {
+      user_timeline: {
+        exclude_replies: false,
+        include_rts: true,
+      },
+    },
+    stream_filters: {
+      replies: false,
+      retweets: false,
+      skipValidAccounts: false,
+    },
   },
 };