+++ src/services/twitter.js
@@ -6,11 +6,12 @@ const pLimit = require('p-limit');
 const uuid = require('uuid/v4');
 const { HttpStatusError } = require('common-errors');
 const {
-  isObject, isString, conforms, merge, find,
+  isObject, isString, conforms, merge, find, isNil,
 } = require('lodash');
 
 const Notifier = require('./notifier');
 const { transform, TYPE_TWEET } = require('../utils/response');
+
 
 function extractAccount(accum, value) {
   const accountId = value.meta.account_id;
@@ -23,6 +24,25 @@ function extractAccount(accum, value) {
   }
 
   return accum;
+}
+
+function twitterApiConfig(config) {
+  const TWITTER_API_DEFAULTS = {
+    // Refer to https://developer.twitter.com/en/docs/twitter-api/v1/tweets/timelines/api-reference/get-statuses-user_timeline
+    user_timeline: {
+      exclude_replies: false,
+      include_rts: true,
+    },
+  };
+  return merge(TWITTER_API_DEFAULTS, config.api);
+}
+
+function streamFilterOptions(config) {
+  const STREAM_FILTERS_DEFAULTS = {
+    replies: false,
+    retweets: false,
+  };
+  return merge({}, STREAM_FILTERS_DEFAULTS, config.stream_filters);
 }
 
 /**
@@ -43,6 +63,28 @@ class Twitter {
     id_str: isString,
     text: isString,
   })
+
+  static isRetweet = (data) => {
+    const retweet = data.retweeted_status;
+    if (isNil(retweet)) {
+      return false;
+    }
+    const tweetOwnerId = get(retweet, 'user.id');
+    // Keep the tweets which are retweeted by the user
+    return tweetOwnerId !== data.user.id;
+  }
+
+  static isReply = (data) => {
+    const toUserId = data.in_reply_to_user_id;
+    if (isNil(toUserId)) {
+      return false;
+    }
+    // Keep the tweets which are replied by the user
+    if (toUserId === data.user.id) {
+      return false;
+    }
+    return !isNil(data.in_reply_to_status_id);
+  }
 
   /**
    * cursor extractor
@@ -95,15 +137,14 @@ class Twitter {
     return tweet;
   }
 
-  static tweetFetcherFactory(twitter, logger) {
+  static tweetFetcherFactory(twitter, logger, apiConfig) {
     const limit = pLimit(1);
     const fetch = (cursor, account, cursorField = 'max_id') => Promise.fromCallback((next) => (
       twitter.get('statuses/user_timeline', {
         count: 200,
         screen_name: account,
         trim_user: false,
-        exclude_replies: false,
-        include_rts: true,
+        ...apiConfig.user_timeline,
         [cursorField]: cursor,
       }, (err, tweets, response) => {
         if (err) {
@@ -153,11 +194,14 @@ class Twitter {
     this.core = core;
     this.client = new TwitterClient(config);
     this.listener = null;
+    this.filterOptions = streamFilterOptions(config);
     this.storage = storage;
     this.logger = logger.child({ namespace: '@social/twitter' });
     this._destroyed = false;
     this.following = [];
-    this.fetchTweets = Twitter.tweetFetcherFactory(this.client, this.logger);
+    this.accountIds = {};
+
+    this.fetchTweets = Twitter.tweetFetcherFactory(this.client, this.logger, twitterApiConfig(config));
 
     // cheaper than bind
     this.onData = (json) => this._onData(json);
@@ -217,6 +261,14 @@ class Twitter {
       : [];
   }
 
+  fillAccountIds(accounts = []) {
+    this.accountIds = accounts.reduce(
+      (map, it) => ({ ...map, [it.account_id]: true }),
+      {}
+    );
+    Object.setPrototypeOf(this.accountIds, null);
+  }
+
   listen(accounts) {
     const params = {};
     if (accounts.length > 0) {
@@ -225,6 +277,7 @@ class Twitter {
         .join(',');
 
       this.setFollowing(accounts);
+      this.fillAccountIds(accounts);
     }
 
     if (!params.follow) {
@@ -325,8 +378,27 @@ class Twitter {
     this._destroyAndReconnect();
   }
 
+  shouldFilterTweet(data) {
+    const { replies, retweets, skipValidAccounts } = this.filterOptions;
+
+    // Don't filter retweets posted by the valid users
+    if (skipValidAccounts && this.accountIds[data.user.id] !== undefined) {
+      return false;
+    }
+    if (replies && Twitter.isReply(data)) {
+      return true;
+    }
+    if (retweets && Twitter.isRetweet(data)) {
+      return true;
+    }
+    return false;
+  }
+
   async _onData(data) {
     if (Twitter.isTweet(data)) {
+      if (this.shouldFilterTweet(data)) {
+        return false;
+      }
       this.logger.debug({ data }, 'inserting tweet');
       try {
         const tweet = Twitter.serializeTweet(data);